fix(traefik): fix Yaegi compatibility for Traefik plugin
The Traefik plugin was failing to load because Yaegi (Traefik's embedded
Go interpreter) cannot handle several Go language features:

1. for range N (Go 1.22+ range-over-integer) causes a nil type panic
   - Changed to traditional for loop in parser.go
   - See: https://github.com/traefik/yaegi/issues/1701

2. math/rand/v2 (Go 1.22+) is not supported by Yaegi
   - Changed to math/rand in ab_ratio.go
   - See: https://github.com/traefik/yaegi/issues/1674

3. syscall package depends on unsafe which Yaegi cannot interpret
   - Split ssrf.go into ssrf.go (common) + ssrf_dialer.go (syscall-based)
   - Dockerfile removes ssrf_dialer.go and provides a stub ssrf_yaegi.go
   - Dial-time SSRF protection unavailable in Traefik; URL-level still works

4. Third-party cache backends (Redis/Memcached) use unsafe via dependencies
   - Dockerfile removes cache_redis and cache_memcached from plugin build
   - Only in-memory cache available in Traefik plugin

Also fixed esi-configuration.yml to use correct test-server port (8080).

Added Yaegi compatibility documentation to servers/traefik/README.md.
