+++ mesi/ssrf.go
@@ -1,13 +1,19 @@
+// SSRF protection for ESI includes.
+//
+// This file contains common SSRF validation functions shared across all
+// server integrations. Dial-time IP blocking (using syscall.RawConn) is
+// in ssrf_dialer.go, which is excluded from Traefik plugin builds because
+// Yaegi cannot interpret the syscall package (it depends on unsafe).
+//
+// See: https://github.com/traefik/yaegi/issues/1636
 package mesi
 
 import (
 	"errors"
 	"fmt"
 	"net"
-	"net/http"
 	"net/url"
 	"strings"
-	"syscall"
 )
 
 var (
@@ -88,41 +94,4 @@ func hostInAllowedHosts(host string, config EsiParserConfig) bool {
 		}
 	}
 	return false
-}
-
-func safeDialer(config EsiParserConfig) *net.Dialer {
-	return &net.Dialer{
-		Control: func(network, address string, c syscall.RawConn) error {
-			host, _, err := net.SplitHostPort(address)
-			if err != nil {
-				return err
-			}
-			ip := net.ParseIP(host)
-			if ip == nil {
-				return fmt.Errorf("internal error: dial address expected to be IP but got: %s", host)
-			}
-			if config.BlockPrivateIPs && isPrivateOrReservedIP(ip) {
-				return fmt.Errorf("%w: blocked dial to private/reserved ip", ErrSSRFBlocked)
-			}
-			return nil
-		},
-	}
-}
-
-// NewSSRFSafeTransport creates an http.Transport that blocks connections to
-// private/reserved IP addresses at dial time, preventing SSRF via DNS rebinding.
-// When config.BlockPrivateIPs is false, it returns a standard transport.
-//
-// Users supplying a custom HTTPClient should use this transport to ensure
-// SSRF protection works correctly:
-//
-//	client := &http.Client{
-//		Transport: mesi.NewSSRFSafeTransport(config),
-//		Timeout:   config.Timeout,
-//	}
-func NewSSRFSafeTransport(config EsiParserConfig) *http.Transport {
-	dialer := safeDialer(config)
-	return &http.Transport{
-		DialContext: dialer.DialContext,
-	}
 }
